requirements-dev.txt
# Development dependencies - NÃO usar em produção

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
pytest-mock>=3.10.0
pytest-xdist>=3.3.0
factory-boy>=3.3.0
coverage>=7.2.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.5.0
pre-commit>=3.3.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Development tools
django-debug-toolbar>=4.1.0
django-extensions>=3.2.0
ipython>=8.14.0
jupyter>=1.0.0

# Database tools (desenvolvimento)
psycopg2-binary>=2.9.0  # PostgreSQL
mysqlclient>=2.2.0      # MySQL

# Performance profiling
django-silk>=5.0.0
memory-profiler>=0.61.0

# API development
djangorestframework>=3.14.0
django-cors-headers>=4.2.0

# File handling
python-magic>=0.4.27